improves handling of spaces and moves phrase styles to separate file
